requirements.txt
PyMuPDF
openai
faiss-cpu
numpy
gradio
tiktoken
